Splitting types.rs. No functional changes.
